Enable usage when running in the application process.
Tweaked which worker the component works through. Ran into an issue if the process that was trying to have a progress bar was running in the application process that it would not work properly.
Now using a worker with a local process name. It is kill when the progress bar quits